hooked up ban user
Admin can ban user from site now. It doesnt allow a banned user to sign
in and if they are signed in then they are logged out on the next page
upload. This is done through the session['status'] in the decideStatus
file. Also fixed the active users div on the chat page.